Create EpisodeBuilder class that queries PRX CMS and includes overrides
The EpisodeBuilder class uses HyperResource to retrieve information from the PRX CMS API. It takes a PRX ID (a new column added to the Episode table) and build a JSON object containing the relevant information about the story that corresponds to that ID. It optionally takes a hash of overrides, which it merges with the CMS data to create the final JSON object. In order to test the EpisodeBuilder class, I have added the WebMock gem and several JSON files as fixtures. The EpisodeBuilder also required setting up the Dotenv gem to store the HAL root.