Change RSS file to respond to a hash instead of an episode object
Instead of simply retrieving the episodes for a given podcast, the podcast controller now also builds a hash of attributes (using the EpisodeBuilder class). This change required adding JSON parsing and symbolize_keys! to the EpisodeBuilder class, because the overrides are stored in the database as a string, not a hash. The podcasts#show view now expects an episode as a hash with symbol keys. I also updated the tests to fit with this change.