Create and test DateUpdater service object
When a podcast's metadata changes, the last build date must be updated, but the pub date can stay the same. The pub date must be updated when new episodes are added, when an episode is edited, and when an episode is deleted. I have not yet dealt with the editing or deleting of episodes because I'm not sure whether those actions will be carried out be administrators or users. When the pub date is updated, the last build date must be updated as well. It looks like there's no getting around using the after_update callback on the Podcast class because editing a podcast is not a route or controller action -- I'm assuming it's something administrators will do through the console.