Validate size and type of podcast images
In the process of implementing FastImage to make sure podcast images comply with iTunes and RSS specifications, I realized that the way I had related images to podcasts was needlessly confusing. Images specifically for iTunes have different requirements and attributes than feed images that aren't meant for iTunes. So I just went ahead and made them separate models. The only quality they have in common is a URL that can't be blank. I added a custom validator for each model that uses FastImage to detect size and type. I'm not exactly sure what the size requirements are for feed images so I just figured they should be smaller than 2048x2048. This change meant that I had to make many minor changes to other tests and factories. I also added a bunch of sample images to test/fixtures for testing purposes. For now, I'm going to ignore episode images, because though it is possible to have them, it seems that very few podcasters do.