Add date calculator given planner params
This commit focuses on the calculate_dates method and subsequent helper
methods. Because of the multiple branching conditionals, periodic vs. monthly
week calculating are separated into different branches, and end conditions
are separated into their own branches. This commit is in the middle of DRYing
up the helper methods in an effort to be more readable and clarify the intent
of the conditional logic. This commit does not render any of the
calculations as it wants to focus on the controller logic first.