Transfer date calculator logic from controller to model
Since I don't want to end up with calculation logic inside the controller,
a PodcastPlanner model was created, and logic moved into that class
proper. The generate_dates! method was created to replace calculate_dates
from the controller, and can now be called on the model instance as needed,
which should also help when I eventually set up Turbo to call this method
as params change.