Add draft generator method to podcast planner
Now that turbo is working, I worked on setting up the draft objects
I will need in order to make an Episode bulk create. This should also
allow me to render the dates from the draft objects themselves, and
sets up a pattern for draft objects to be added/removed/updated
independently.