Add simple calendar gem and draft calendar
After adding the gem, I am learning how the monthly_calendar feature
works and how to implement the current PodcastPlanner class into
using it. In this PR there is a base idea for how the calendars can
be flexibly added/subtracted based on the dates that are generated
using the tool.