Refactor calendar rendering not based on preselected dates
The calendars rendering is now decoupled from needing any preselected
dates present. This simplified the need to do any date sorting by
month in the model or needing to set that up in the controller, and
we can now just pass in a set of dates to check against for
pre-selection.